Same thing for Writer
- Make ExtensionManager a decorator
